+++ src/i18next.js
@@ -117,10 +117,14 @@ class I18n extends EventEmitter {
         simplifyPluralSuffix: this.options.simplifyPluralSuffix,
       });
 
+      const usingLegacyFormatFunction = this.options.interpolation.format && this.options.interpolation.format !== defOpts.interpolation.format;
+      if (usingLegacyFormatFunction) {
+        this.logger.warn(`init: you are still using the legacy format function, please use the new approach: https://www.i18next.com/translation-function/formatting`);
+      }
+
       if (formatter && (!this.options.interpolation.format || this.options.interpolation.format === defOpts.interpolation.format)) {
         s.formatter = createClassOnDemand(formatter);
-        s.formatter.init(s, this.options);
-
+        if (s.formatter.init) s.formatter.init(s, this.options);
         this.options.interpolation.format = s.formatter.format.bind(s.formatter);
       }
 
@@ -532,13 +536,13 @@ class I18n extends EventEmitter {
     if (!lng) lng = this.resolvedLanguage || (this.languages?.length > 0 ? this.languages[0] : this.language);
     if (!lng) return 'rtl';
 
-    if (Intl.Locale) {
+    try {
       const l = new Intl.Locale(lng)
       if (l && l.getTextInfo) {
         const ti = l.getTextInfo()
         if (ti && ti.direction) return ti.direction
       }
-    }
+    } catch (e) {/* fall through */}
 
     const rtlLngs = [
       'ar',